modified login api and verify api and validations and add status code

+++ authentication/views.py
@@ -26,6 +26,8 @@ import random
 from rest_framework.permissions import AllowAny,IsAuthenticated
 from rest_framework.authentication import TokenAuthentication
 from rest_framework import exceptions
+from rest_framework import status
+
 
 class BearerTokenAuthentication(TokenAuthentication):
     keyword = 'Bearer'
@@ -49,69 +51,25 @@ class BearerTokenAuthentication(TokenAuthentication):
 def test(request):
     return render(request,'authentication/test.html')
 
-
 class RegisterApi(APIView):
     permission_classes = [AllowAny]
 
     def post(self, request, format=None):
-        # Get data from request
-        name = request.data.get('name')
-        gender = request.data.get('gender')
-        phone_number = request.data.get('phone')
-        mohalla = request.data.get('mohalla')
-        village = request.data.get('village')
-        user_type = request.data.get('user_type')
-        state = request.data.get('state')
-        district = request.data.get('district')
-        latitude= request.data.get('latitude')
-        longitude=request.data.get('longitude')
+        serializer = RegisterSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
 
-
-        # Validate phone number
-        if not phone_number:
-            return Response({'error': 'Phone number is required.'})
-        if not re.match(r'^\d{10}$', phone_number):
-            return Response({'error': 'Phone number should be 10 digits.'})
-        # Validate user_type
-        if not user_type:
-            return Response({'error': 'User type is required.'})
-        if user_type not in ["Grahak", "Sahayak", "MachineMalik"]:
-            return Response({'error': 'User type should be one of the following: Grahak, Sahayak, MachineMalik.'})
-
-        # Validate other fields
-        if not name:
-            return Response({'error': 'Name is required.'})
-        if not gender:
-            return Response({'error': 'Gender is required.'})
-        if gender not in ["Male", "Female"]:
-            return Response({'error': 'Gender should be Male or Female'})    
-        if not mohalla:
-            return Response({'error': 'Mohalla is required.'})
-        if not village:
-            return Response({'error': 'Village is required.'})
-        if not state:
-            return Response({'error': 'State is required.'})
-        if not district:
-            return Response({'error': 'District is required.'})
-        if not latitude or not longitude:
-            return Response({'error':'latitude and longitude can not null'})
-        if not re.match(r'^[-+]?([1-8]?\d(\.\d+)?|90(\.0+)?)$', str(latitude)):
-            return Response({'error': 'Invalid latitude format.'})
-        if not re.match(r'^[-+]?(180(\.0+)?|((1[0-7]\d)|([1-9]?\d))(\.\d+)?)$', str(longitude)):
-            return Response({'error': 'Invalid longitude format.'})
-        try:
-            state = State.objects.get(name=state)
-        except State.DoesNotExist:
-            return Response({'error': f'State with name {state} does not exist in the database.'})
-        
-        # Check if district exists within the state
-        try:
-            district = District.objects.get(name=district, state=state)
-        except District.DoesNotExist:
-            return Response({'error': f'District with name {district} does not exist within the state {state}.'})
-        # Check if user already exists
-        if User.objects.filter(mobile_no=phone_number).exists():
-            return Response({'error': 'User already exists'})
+        # Get validated data from serializer
+        validated_data = serializer.validated_data
+        name = validated_data['name']
+        gender = validated_data['gender']
+        phone_number = validated_data['phone']
+        mohalla = validated_data['mohalla']
+        village = validated_data['village']
+        user_type = validated_data['user_type']
+        state = validated_data['state']
+        district = validated_data['district']
+        latitude = validated_data['latitude']
+        longitude = validated_data['longitude']
 
         # Generate OTP and create user, profile, and OTP objects
         otp = random.randint(100000, 999999)
@@ -137,7 +95,7 @@ class RegisterApi(APIView):
             otp=otp,
             user=user
         )
-        
+
         # Send OTP to user's phone number (use your own SMS gateway or API)
         # TODO: Implement SMS gateway/API integration to send OTP to user's phone number
         # ...
@@ -146,32 +104,44 @@ class RegisterApi(APIView):
         return Response({
             'success': True,
             'otp': otp_obj.otp,
+            'status':status.HTTP_201_CREATED
         })
     
-
 class VerifyMobile(APIView):
     permission_classes=[AllowAny,]
-
     def post(self, request, format=None):
         otp_input = request.data.get('otp')
-        try:
-            otp = OTP.objects.get(otp=otp_input)
-            user = otp.user
-        except OTP.DoesNotExist:
-            return Response({'error': 'Invalid OTP'})
-
-        # Create new token for the user
-        use=Token.objects.filter(user=user)
-        print(use)
+        phone=request.data.get('phone')
+        if not otp_input or not phone:
+            return Response({'error':'otp and phone both is required !'})
+        if not re.match(r'^\d{10}$', phone):
+            return Response({'error': 'Phone number should be 10 digits.'})
         # Revoke any existing tokens for the user
-        Token.objects.filter(user=user).delete()
-        token, _ = Token.objects.get_or_create(user=user)
+        user = authenticate(request, mobile_no=phone)
+        if user is not None:
+            # get all sessions of the user and delete them
+            sessions = Session.objects.filter(expire_date__gte=timezone.now(), session_key__contains=str(user.id))
+            for session in sessions:
+                session.delete()
+            # delete all tokens of the user
+            Token.objects.filter(user=user).delete()
+            # authenticate and login the user
+            login(request, user)
+        try:
+            get_user=User.objects.get(mobile_no=phone)
+            otp = OTP.objects.get(otp=otp_input,user=get_user)
+            user_get = otp.user
+        except User.DoesNotExist:
+            return Response({'error':'User does not exist !'})    
+        except OTP.DoesNotExist:
+            return Response({'error': 'Invalid OTP'})    
+        token, _ = Token.objects.get_or_create(user=user_get)
         # Update user properties
-        user.is_verified = True
-        user.is_active = True
-        user.save()
-        OTP.objects.filter(user=user).delete()
-        return Response({'verified':True,'token': token.key})
+        user_get.is_verified = True
+        user_get.is_active = True
+        user_get.save()
+        OTP.objects.filter(user=user_get).delete()
+        return Response({'verified':True,'token': token.key,'status':status.HTTP_200_OK})
     
 @api_view(["POST"])
 @authentication_classes([BearerTokenAuthentication])
@@ -181,43 +151,30 @@ def User_logout(request):
     logout(request)
     response_data = {
         'msg': 'user logged out successfully !',
-        'msg2':'previous token has been deleted !'
+        'msg2':'previous token has been deleted !',
+        'status':status.HTTP_200_OK
+
     }
     return Response(response_data)
-
 
 class LoginApi(APIView):
     permission_classes = [AllowAny,]
     @method_decorator(never_cache)
     def post(self, request, *args, **kwargs):
-        phone = request.data.get('phone')
-        if not phone:
-            return Response({'error':'enter valid phone number !'})
-        if not re.match(r'^\d{10}$', phone):
-            return Response({'error': 'Phone number should be 10 digits.'})
-        
+        serializer = LoginSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        phone = serializer.validated_data['phone']
         user = authenticate(request, mobile_no=phone)
-        if user is not None:
-            # get all sessions of the user and delete them
-            sessions = Session.objects.filter(expire_date__gte=timezone.now(), session_key__contains=str(user.id))
-            for session in sessions:
-                session.delete()
-
-            # delete all tokens of the user
-            Token.objects.filter(user=user).delete()
-
-            # authenticate and login the user
-            login(request, user)
-
+        if user is not None:           
             # create OTP and send it to the user
             otp = random.randint(100000, 999999)
             otps = OTP.objects.create(otp=otp, user=user)
             # send OTP to the user's mobile number for verification
 
-            return Response({'success':True,'otp':otps.otp,'user_type':user.user_type,'data':'Previous sessions and tokens deleted successfully'})
+            return Response({'success': True, 'otp': otps.otp, 'user_type': user.user_type, 'data': 'Previous sessions and tokens deleted successfully', 'status': status.HTTP_200_OK})
         else:
-            return Response({'error': 'Invalid phone number or user not registered'})
-
+            return Response({'message': 'User not registered !','status':status.HTTP_404_NOT_FOUND})
+        
 class MobileNoAuthBackend(BaseBackend):
     def authenticate(self, request, mobile_no=None):
         try:
@@ -241,44 +198,42 @@ class IsOwnerOrReadOnly(BasePermission):
     def has_object_permission(self, request, view, obj):
         return obj.user == request.user or request.method in ['GET','POST','PUT', 'HEAD', 'OPTIONS']        
 ###-------------------------------------------------------------------------------####
-class chartJsonListView(View):
-    def get(self, *args, **kwargs):
-        today = date.today()
-        vendor = Profile.objects.get(user=self.request.user)
+# class chartJsonListView(View):
+#     def get(self, *args, **kwargs):
+#         today = date.today()
+#         vendor = Profile.objects.get(user=self.request.user)
 
-        product_count_list = []
-        order_count_list = []
-        # for i in range(1, 13):
-        #     product_count = Product.objects.filter(
-        #         product_vendor=vendor,  date__year=today.year, date__month=i,).count()
-        #     product_count_list.append(product_count)
-        #     order_count = OrderSupplier.objects.all().filter(vendor=vendor, order_date__year=today.year,
-        #                                                         order_date__month=i,).exclude(status="PENDING").count()
-        #     order_count_list.append(order_count)
+#         product_count_list = []
+#         order_count_list = []
+#         # for i in range(1, 13):
+#         #     product_count = Product.objects.filter(
+#         #         product_vendor=vendor,  date__year=today.year, date__month=i,).count()
+#         #     product_count_list.append(product_count)
+#         #     order_count = OrderSupplier.objects.all().filter(vendor=vendor, order_date__year=today.year,
+#         #                                                         order_date__month=i,).exclude(status="PENDING").count()
+#         #     order_count_list.append(order_count)
 
-        return JsonResponse({"product_count_list": 12, "order_count_list": 12, }, safe=False)
+#         return JsonResponse({"product_count_list": 12, "order_count_list": 12, }, safe=False)
 
 
-class chartJsonListViewAdmin(View):
-    def get(self, *args, **kwargs):
-        today = date.today()
-        user = User.objects.get(username=self.request.user.username)
-        if user.is_superuser == True:
-            # vendor = Profile.objects.get(user=self.request.user)
-            product_count_list = []
-            order_count_list = []
-            # for i in range(1, 13):
-            #     product_count = Product.objects.all().filter(
-            #         date__year=today.year, date__month=i,).count()
-            #     product_count_list.append(product_count)
-            #     order_count = OrderSupplier.objects.all().filter(order_date__year=today.year,
-            #                                                         order_date__month=i,).exclude(status="PENDING").count()
-            #     order_count_list.append(order_count)
+# class chartJsonListViewAdmin(View):
+#     def get(self, *args, **kwargs):
+#         today = date.today()
+#         user = User.objects.get(username=self.request.user.username)
+#         if user.is_superuser == True:
+#             # vendor = Profile.objects.get(user=self.request.user)
+#             product_count_list = []
+#             order_count_list = []
+#             # for i in range(1, 13):
+#             #     product_count = Product.objects.all().filter(
+#             #         date__year=today.year, date__month=i,).count()
+#             #     product_count_list.append(product_count)
+#             #     order_count = OrderSupplier.objects.all().filter(order_date__year=today.year,
+#             #                                                         order_date__month=i,).exclude(status="PENDING").count()
+#             #     order_count_list.append(order_count)
 
-            return JsonResponse({"product_count_list": 12, "order_count_list": 12, }, safe=False)
+#             return JsonResponse({"product_count_list": 12, "order_count_list": 12, }, safe=False)
 ###-------------------------------------------------------------------------------------------####
-
-
